refactor: Organize Auction domain
