+++ frontend/src/features/auction/components/mockData.ts
