+++ frontend/src/features/auction/types/AuctionItem.ts
