+++ frontend/src/features/auction/types/AuctionItemProps.ts
