+++ frontend/src/features/auction/types/AuctionListProps.ts
