+++ frontend/src/features/auction/types/AuctionModalProps.ts
