+++ frontend/src/features/auction/types/AuctionSearchProps.ts
